Refactor : 사용하지 않는 코드 삭제
사용하지 않는 코드들을 삭제합니다.
