+++ src/main/java/nbbang/com/nbbang/domain/chat/controller/ChatResponseMessage.java
@@ -2,18 +2,13 @@ package nbbang.com.nbbang.domain.chat.controller;
 
 public class ChatResponseMessage {
     public static final String READ_CHAT = "채팅방 조회 성공";
-    public static final String UPDATE_CHAT = "채팅방 정보 업데이트 성공";
 
     public static final String UPLOADED_MESSAGE = "채팅방 메시지 전송 성공";
     public static final String UPLOADED_IMAGE_MESSAGE = "채팅방 메시지 이미지 전송 성공";
 
-    public static final String ILLEGAL_CHAT_STATUS = "채팅방 상태 정보를 올바르게 입력하세요.";
-    public static final String ILLEGAL_CHAT_GOAL_NUMBER = "채팅방 최대 참여자 수를 올바르게 입력하세요.";
     public static final String ILLEGAL_CHAT_EMPTY= "채팅 메시지는 비어있을 수 없습니다.";
     public static final String ILLEGAL_CHAT_LENGTH = "채팅 메시지의 길이는 1000자 이하이어야 합니다.";
 
     public static final String MESSAGE_NOT_FOUND = "존재하지 않는 메시지입니다.";
-
-    public static final String MESSAGE_READ_OK = "메시지 읽음을 잘 반영하였습니다.";
 
 }
